# C4-262 Fix main page inefficiency