--- a/chalicelib/app_utils.py
+++ b/chalicelib/app_utils.py
@@ -236,8 +236,32 @@ def query_params_to_literals(params):
     return params
 
 
+def get_size(obj, seen=None):
+    """ Recursively finds size of objects
+        Taken directly from: https://goshippo.com/blog/measure-real-size-any-python-object/
+    """
+    size = sys.getsizeof(obj)
+    if seen is None:
+        seen = set()
+    obj_id = id(obj)
+    if obj_id in seen:
+        return 0
+    # Important mark as seen *before* entering recursion to gracefully handle
+    # self-referential objects
+    seen.add(obj_id)
+    if isinstance(obj, dict):
+        size += sum([get_size(v, seen) for v in obj.values()])
+        size += sum([get_size(k, seen) for k in obj.keys()])
+    elif hasattr(obj, '__dict__'):
+        size += get_size(obj.__dict__, seen)
+    elif hasattr(obj, '__iter__') and not isinstance(obj, (str, bytes, bytearray)):
+        size += sum([get_size(i, seen) for i in obj])
+    return size
+
+
 def trim_output(output, max_size=100000):
-    """ Uses sys.getsizeof instead of encoding as JSON to determine if the output size is too large.
+    """ Uses the helper above with sys.getsizeof to determine the output size and remove it if it is too large.
+        Instead of encoding as JSON as that is very slow.
 
     Old docstring below:
 
@@ -254,7 +278,7 @@ def trim_output(output, max_size=100000):
     #     return ''.join([formatted[:max_size], '\n\n... Output truncated ...'])
     # else:
     #     return formatted
-    size = sys.getsizeof(output)
+    size = get_size(output)
     if size > max_size:
         return 'Output too large to provide on main page - see check result directly'
     return output