- When getting the main page, for all checks that have an action we (synchronously) attempt to get the latest result for that action and populate `latest_action` with the action name.
- This causes the main page to take several seconds longer to load
- Instead, show the action history link for all checks by default (without the additional get per check). This does mean the link will show up for checks that do not have any action history, but I think that is acceptable given the cost.
- Reduces page load time from ~18 seconds to ~12 seconds on local (approximately ~7 to ~5 seconds on production)
- Optimizes out some unneeded API calls
- Optimizes out lots of calls to `json.dumps` while trimming output, which is very slow, instead favoring using a utility based off of `sys.getsizeof` which is faster.